+++ src/tables/mod.rs
@@ -10,6 +10,7 @@ pub mod base;
 pub mod cbdt;
 pub mod cblc;
 pub mod cff;
+pub mod cff2;
 pub mod cmap;
 pub mod colr;
 pub mod cpal;